(https://github.com/mas2020-python/pycryptex/issues/13) keep the --remove option default in encrypt and decrypt

+++ tests/test_encrypt.py
@@ -5,12 +5,17 @@ from pycryptex.main import cli
 
 def test_encrypt():
     runner = CliRunner()
-    os.remove('test/secrets.txt.enc')
+    if os.path.exists('test/secrets.txt.enc'):
+        os.remove('test/secrets.txt.enc')
     result = runner.invoke(cli,
-                           ['encrypt', '--pubkey', 'test/id_rsa.pub', 'test/secrets.txt'])
+                           ['encrypt', '--pubkey', 'test/id_rsa.pub', 'test/secrets.txt', '--keep'])
     assert result.exit_code == 0
     assert not result.exception
     assert os.path.exists('test/secrets.txt.enc') is True
+    result = runner.invoke(cli,
+                           ['encrypt', '--pubkey', 'test/id_rsa.pub', 'test/secrets.txt'])
+    assert os.path.exists('test/secrets.txt.enc') is True
+    assert os.path.exists('test/secrets.txt') is False
 
 
 def test_decrypt():
